Chamado dos metodos corba no ambiente RMI e metodos corba e RMI no ws

+++ SD_PROJETO_FINAL/src/rmi/RmiFunctions.java
@@ -4,10 +4,10 @@ import java.rmi.*;
 
 public interface RmiFunctions extends Remote {
 	
-   public boolean reservarRestaurante(int idRestaurante, int qtdPessoas) throws RemoteException;
+   public boolean reservarRestaurante(short idRestaurante, int qtdPessoas) throws RemoteException;
    
-   public boolean comprarIngressoCinema(int idCinema, int idFilme, int horario, int qtdIngressos) throws RemoteException;
+   public boolean comprarIngressoCinema(short idCinema, short idFilme, short horario, int qtdIngressos) throws RemoteException;
    
-   public boolean comprarIngressoTeatro(int idTeatro, int idPeca, int idHorario, int qtdIngressos) throws RemoteException;
+   public boolean comprarIngressoTeatro(short idTeatro, short idPeca, short idHorario, int qtdIngressos) throws RemoteException;
    
 }
